+++ packages/server/src/api/database/index.js
@@ -1,5 +1,7 @@
 const log = require('@home-gallery/logger')('server.api.database');
 
+const { filterEntriesByQuery } = require('@home-gallery/query')
+
 const { waitReadWatch } = require('./read-database');
 const { cache } = require('./cache-middleware');
 const { sanitizeInt } = require('./sanitize');
@@ -17,6 +19,8 @@ function uniqEntries(entries) {
   return uniqEntries;
 }
 
+const filterDatabase = (entries, query, cb) => query ? filterEntriesByQuery(entries, query, cb) : cb(null, entries)
+
 function databaseApi(eventbus) {
   let database = false;
   const databaseCache = cache(3600);
@@ -32,12 +36,22 @@ function databaseApi(eventbus) {
         }
       }
       return res.status(404).json(err);
-    } else if (req.query && (req.query.offset || req.query.limit)) {
-      const length = database.data.length;
-      const offset = sanitizeInt(req.query.offset, 0, length, 0);
-      const limit = sanitizeInt(req.query.limit, Math.min(10, length), length, length);
-      const data = database.data.slice(offset, offset + limit);
-      res.send(Object.assign({}, database, { limit, offset, data }));
+    } else if (req.query && (req.query.offset || req.query.limit || req.query.q)) {
+      filterDatabase(database.data, req.query.q, (err, entries) => {
+        if (err) {
+          log.error(err, `Failed to filter database with query '${req.query.q}': ${err}`)
+          return res.status(400).json({
+            type: 'error',
+            code: 400,
+            message: `Failed to filter database with query '${req.query.q}': ${err}`
+          })
+        }
+        const length = entries.length;
+        const offset = sanitizeInt(req.query.offset, 0, length, 0);
+        const limit = sanitizeInt(req.query.limit, Math.min(10, length), length, length);
+        const data = entries.slice(offset, offset + limit);
+        res.send(Object.assign({}, database, { limit, offset, data }));
+      })
     } else {
       res.send(database);
     }